project_2/requirements.txt
pandas
scikit-learn
numpy
mlflow
azureml-sdk
joblib
